requirements.txt
customtkinter
opencv-python
face_recognition
numpy
pandas
scikit-learn
pillow